backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "asgi-lifespan==2.*",
    "configargparse>=1.7",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "psycopg[binary]>=3.2.6",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.5",
    "sqlalchemy-utils>=0.41.2",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
explicit_package_bases = true
strict = true
check_untyped_defs = true
exclude = [".*/tests.*"]
